+++ selfservice/flow/registration/error.go
@@ -1,9 +1,13 @@
 package registration
 
 import (
+	"context"
 	"fmt"
+	"github.com/ory/kratos/selfservice/form"
+	"github.com/ory/x/errorsx"
 	"net/http"
 	"net/url"
+	"time"
 
 	"github.com/pkg/errors"
 
@@ -18,10 +22,6 @@ import (
 
 var (
 	ErrHookAbortRequest = errors.New("abort hook")
-
-	ErrRequestExpired = herodot.ErrBadRequest.
-				WithError("registration request expired").
-				WithReasonf(`The registration request has expired. Please restart the flow.`)
 )
 
 type (
@@ -31,6 +31,7 @@ type (
 		x.LoggingProvider
 
 		RequestPersistenceProvider
+		HandlerProvider
 	}
 
 	ErrorHandlerProvider interface{ RegistrationRequestErrorHandler() *ErrorHandler }
@@ -39,7 +40,20 @@ type (
 		d errorHandlerDependencies
 		c configuration.Provider
 	}
+
+	requestExpiredError struct {
+		*herodot.DefaultError
+	}
 )
+
+func newRequestExpiredError(since time.Duration) requestExpiredError {
+	return requestExpiredError{
+		herodot.ErrBadRequest.
+			WithError("registration request expired").
+			WithReasonf(`The registration request has expired. Please restart the flow.`).
+			WithReasonf("The registration request expired %.2f minutes ago, please try again.", since.Minutes()),
+	}
+}
 
 func NewErrorHandler(d errorHandlerDependencies, c configuration.Provider) *ErrorHandler {
 	return &ErrorHandler{
@@ -60,6 +74,25 @@ func (s *ErrorHandler) HandleRegistrationError(
 		WithField("credentials_type", ct).
 		WithField("login_request", rr).
 		Warn("Encountered login error.")
+
+	if _, ok := errorsx.Cause(err).(requestExpiredError); ok {
+		// create new request because the old one is not valid
+		if err = s.d.RegistrationHandler().NewRegistrationRequest(w, r, func(a *Request) (string, error) {
+			for name, method := range a.Methods {
+				method.Config.AddError(&form.Error{Message: "Your session expired, please try again."})
+				if err := s.d.RegistrationRequestPersister().UpdateRegistrationRequest(context.TODO(), a.ID, name, method); err != nil {
+					return s.d.SelfServiceErrorManager().Create(r.Context(), w, r, err)
+				}
+				a.Methods[name] = method
+			}
+
+			return urlx.CopyWithQuery(s.c.RegisterURL(), url.Values{"request": {a.ID.String()}}).String(), nil
+		}); err != nil {
+			// failed to create a new session and redirect to it, handle that error as a new one
+			s.HandleRegistrationError(w, r, ct, rr, err)
+		}
+		return
+	}
 
 	if rr == nil {
 		s.d.SelfServiceErrorManager().Forward(r.Context(), w, r, err)