+++ selfservice/flow/registration/request.go
@@ -19,25 +19,37 @@ import (
 type Request struct {
 	// ID represents the request's unique ID. When performing the registration flow, this
 	// represents the id in the registration ui's query parameter: http://<urls.registration_ui>/?request=<id>
+	//
+	// required: true
 	ID uuid.UUID `json:"id" faker:"uuid" db:"id" rw:"r"`
 
 	// ExpiresAt is the time (UTC) when the request expires. If the user still wishes to log in,
 	// a new request has to be initiated.
+	//
+	// required: true
 	ExpiresAt time.Time `json:"expires_at" faker:"time_type" db:"expires_at"`
 
 	// IssuedAt is the time (UTC) when the request occurred.
+	//
+	// required: true
 	IssuedAt time.Time `json:"issued_at" faker:"time_type" db:"issued_at"`
 
 	// RequestURL is the initial URL that was requested from ORY Kratos. It can be used
 	// to forward information contained in the URL's path or query for example.
+	//
+	// required: true
 	RequestURL string `json:"request_url" db:"request_url"`
 
 	// Active, if set, contains the registration method that is being used. It is initially
 	// not set.
+	//
+	// required: true
 	Active identity.CredentialsType `json:"active,omitempty" db:"active_method"`
 
 	// Methods contains context for all enabled registration methods. If a registration request has been
 	// processed, but for example the password is incorrect, this will contain error messages.
+	//
+	// required: true
 	Methods map[identity.CredentialsType]*RequestMethod `json:"methods" faker:"registration_request_methods" db:"-"`
 
 	// MethodsRaw is a helper struct field for gobuffalo.pop.
@@ -112,7 +124,7 @@ func (r *Request) GetID() uuid.UUID {
 
 func (r *Request) Valid() error {
 	if r.ExpiresAt.Before(time.Now()) {
-		return errors.WithStack(ErrRequestExpired.WithReasonf("The registration request expired %.2f minutes ago, please try again.", time.Since(r.ExpiresAt).Minutes()))
+		return errors.WithStack(newRequestExpiredError(time.Since(r.ExpiresAt)))
 	}
 	if r.IssuedAt.After(time.Now()) {
 		return errors.WithStack(herodot.ErrBadRequest.WithReason("The registration request was issued in the future."))