+++ selfservice/form/fields.go
@@ -10,6 +10,8 @@ import (
 	"github.com/ory/kratos/schema"
 )
 
+const DisableFormField = "disableFormField"
+
 // Fields contains multiple fields
 //
 // swagger:model formFields
@@ -20,17 +22,28 @@ type Fields []Field
 // swagger:model formField
 type Field struct {
 	// Name is the equivalent of <input name="{{.Name}}">
+	//
+	// required: true
 	Name string `json:"name"`
 	// Type is the equivalent of <input type="{{.Type}}">
-	Type string `json:"type,omitempty"`
+	//
+	// required: true
+	Type string `json:"type"`
+
 	// Pattern is the equivalent of <input pattern="{{.Pattern}}">
 	Pattern string `json:"pattern,omitempty"`
+
 	// Disabled is the equivalent of <input disabled="{{.Disabled}}">
-	Disabled string `json:"disabled,omitempty"`
+	Disabled bool `json:"disabled,omitempty"`
+
 	// Required is the equivalent of <input required="{{.Required}}">
+	//
+	// required: true
 	Required bool `json:"required,omitempty"`
+
 	// Value is the equivalent of <input value="{{.Value}}">
 	Value interface{} `json:"value,omitempty" faker:"name"`
+
 	// Errors contains all validation errors this particular field has caused.
 	Errors []Error `json:"errors,omitempty"`
 }
@@ -113,6 +126,13 @@ func fieldFromPath(name string, p jsonschemax.Path) Field {
 	// Other properties
 	if p.Pattern != nil {
 		f.Pattern = p.Pattern.String()
+	}
+
+	// Set disabled if the custom property is set
+	if isDisabled, ok := p.CustomProperties[DisableFormField]; ok {
+		if isDisabled, ok := isDisabled.(bool); ok {
+			f.Disabled = isDisabled
+		}
 	}
 
 	return f